requirements.txt
streamlit
Pillow
transformers
torch
